+++ lib/common/taper.c
@@ -21,7 +21,6 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <types.h>
-#include <common/memory.h>
 #include <cgraph/alloc.h>
 #include <cgraph/agxbuf.h>
 #include <cgraph/list.h>