+++ lib/gvc/gvjobs.c
@@ -11,7 +11,6 @@
 #include "config.h"
 
 #include	<cgraph/alloc.h>
-#include	<common/memory.h>
 #include	<common/types.h>
 #include        <gvc/gvplugin.h>
 #include        <gvc/gvcjob.h>
@@ -102,11 +101,10 @@ GVJ_t *gvjobs_next(GVC_t * gvc)
     return (gvc->job = job);
 }
 
-void gv_argvlist_set_item(gv_argvlist_t *list, int index, char *item)
-{
+void gv_argvlist_set_item(gv_argvlist_t *list, size_t index, char *item) {
     if (index >= list->alloc) {
+	list->argv = gv_recalloc(list->argv, list->alloc, index + 10, sizeof(char *));
 	list->alloc = index + 10;
-	list->argv = grealloc(list->argv, (list->alloc)*(sizeof(char*)));
     }
     list->argv[index] = item;
 }