+++ lib/neatogen/fPQ.h
@@ -9,7 +9,6 @@
  *************************************************************************/
 
 #include <cgraph/alloc.h>
-#include <common/memory.h>
 #include <stdio.h>
 
 /* Priority queue: